Fix a white screen related to DONOTCACHEPAGE.
When DONOTCACHEPAGE is defined by a plugin (e.g. in an init hook), Quick Cache
serves an empty page. This happens because:
At startup, if DONOTCACHEPAGE isn't defined, an object buffer is created.
At shutdown, if DONOTCACHEPAGE isn't defined, this same buffer is printed.
If DONOTCACHEPAGE isn't defined at startup, but is defined at shutdown,
then the page's contents are in a buffer, but that buffer isn't being printed.

The fix defines DONOTCACHEPAGE to false, if it's not already defined.
This stops other plugins from defining it to true between startup and shutdown.