If WP_TEMP_DIR is set use this
When Zencache tries to guess the path to the tmp-diretory) on IIS Win2008R2 with strict open_basedir() enabled, it returns a non-writable directory. It seems the PHP function is_writables() is reporting the directory it tries first to be writable even it's outside the open_basedir().

Tried to manually set the dir with WP_TEMP_DIR, but since it's guessing on the other possible directories it does

If a user explicit sets WP_TEMP_DIR I think it should be the path used.